Changed stakes to tsx
